pyproject.toml
[project]
name = "socks5mitm"
version = "v0.0.0"
description = " This library implements a basic SOCKS5 server with MITM features."
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["mitm", "socks5", "mitmproxy", "socks5-server", "mitm-server", "socks5-proxy-server"]

dependencies = ["pysocks"]

[project.optional-dependencies]
dev = ["ruff", "pytest", "mypy"]

[build-system]
requires = ["setuptools>=43.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["src"]
